unified-report-service/requirements.txt
fastapi
uvicorn
pydantic
reportlab
jinja2
cvss
requests
python-dotenv